Moved the windows gst_extractor module to portable.
Made linux use this one as well.  The only real difference between the two was
the code to add timeouts, which I just forgot to add to linux when I added it
to windows.
