+++ tv/linux/plat/utils.py
@@ -311,7 +311,7 @@ def exit_miro(return_code):
     sys.exit(return_code)
 
 def run_media_metadata_extractor(movie_path, thumbnail_path):
-    from miro.plat.renderers import gst_extractor
+    from miro.frontends.widgets.gst import gst_extractor
     return gst_extractor.run(movie_path, thumbnail_path)
 
 def miro_helper_program_info():
